/deep/
That is another nice special, Selector, which allows you to apply stars not just to the elements in a template. So not just to the elements inside of the components, but also to the same elements in a child component of that component to see how deep works.

So even if this child component then had another child component, both child components would receive this. So it overrides it for the whole chain of child components, not just for the direct one.

/deep/ & >>> Selectors allow you to override styles of a specific element or of course, any other selector. So combining an element with a class or having just a class override that not only in the component itself, but also all child components.

Super helpful in some cases.